workerbee/requirements.txt
cassandra-driver
git+https://github.com/waggle-sensor/pywaggle
numpy
pandas
jinja2
